requirements.txt
spacy~=3.6.1
mysql~=0.0.3
mysql-connector-python~=8.1.0
setuptools~=68.0.0
gensim~=4.3.1
spacytextblob~=4.0.0
pyenchant~=3.2.2
